feat: write the view of user operation

+++ HP-app/src/main/java/com/graduation/hp/presenter/UserTabPresenter.java
@@ -7,13 +7,11 @@ import com.graduation.hp.core.repository.http.bean.Result;
 import com.graduation.hp.repository.RepositoryHelper;
 import com.graduation.hp.repository.contact.UserTabContact;
 import com.graduation.hp.repository.model.impl.UserModel;
-import com.graduation.hp.ui.navigation.center.UserTabFragment;
+import com.graduation.hp.ui.navigation.user.UserTabFragment;
 
 import javax.inject.Inject;
 
 import io.reactivex.Single;
-import io.reactivex.SingleEmitter;
-import io.reactivex.SingleOnSubscribe;
 import io.reactivex.functions.Consumer;
 
 public class UserTabPresenter extends BasePresenter<UserTabFragment, UserModel>
@@ -31,7 +29,7 @@ public class UserTabPresenter extends BasePresenter<UserTabFragment, UserModel>
             String icon = repositoryHelper.getPreferencesHelper().getCurrentUserIcon();
             String nickname = repositoryHelper.getPreferencesHelper().getCurrentUserNickname();
             float bmi = repositoryHelper.getPreferencesHelper().getCurrentUserBMI();
-            int healthy = repositoryHelper.getPreferencesHelper().getCurrentUserHealthyNum();
+            long healthy = repositoryHelper.getPreferencesHelper().getCurrentUserHealthyNum();
             boolean isCurrentUserLogin = !TextUtils.isEmpty(repositoryHelper.getPreferencesHelper().getCurrentUserToken());
             mMvpView.getCurrentUserInfoSuccess(isCurrentUserLogin, icon, nickname, bmi, healthy);
         }));
