+++ HP-app/src/main/java/com/graduation/hp/repository/contact/UserTabContact.java
@@ -4,7 +4,7 @@ public interface UserTabContact {
 
     interface View {
         void getCurrentUserInfoSuccess(boolean isCurrentUserLogin, String icon,
-                                       String nickname, float bmi, int healthyNum);
+                                       String nickname, float bmi, long healthyNum);
     }
 
     interface Presenter {
